Created a new MemoryMatch Scene and SoundManager
Created a new memory match game manager and scene as this one has more functionality. The pots now get highlighted when being selected and while correctly matched. The pots will also turn over automatically if the matches are incorrect or the game is restarting.

The SoundManager was created and can be used throughout the entire game.